Cleanup headers of libYARP_robotinterface
* Fix include guards  of RobotInterfaceDTD
* Make XMLReaderFileVx, XMLReaderFileV1 and XMLReaderFileV3 private classes
