+++ src/libYARP_robotinterface/src/yarp/robotinterface/XMLReader.h
@@ -17,41 +17,7 @@ namespace yarp {
 namespace robotinterface {
 
 class Robot;
-
-
-class YARP_robotinterface_API XMLReaderFileVx
-{
-public:
-    bool verbose;
-    virtual ~XMLReaderFileVx() {};
-    virtual  Robot& getRobotFile(const std::string &filename, bool verbose = false)=0;
-};
-
-class YARP_robotinterface_API XMLReaderFileV1 : public XMLReaderFileVx
-{
-public:
-    XMLReaderFileV1();
-    virtual ~XMLReaderFileV1();
-
-    Robot& getRobotFile(const std::string &filename, bool verbose = false) override;
-
-private:
-    class privateXMLReaderFileV1;
-    privateXMLReaderFileV1 * const mPriv;
-};
-
-class YARP_robotinterface_API XMLReaderFileV3: public XMLReaderFileVx
-{
-public:
-    XMLReaderFileV3();
-    virtual ~XMLReaderFileV3();
-
-    Robot& getRobotFile(const std::string &filename, bool verbose = false) override;
-
-private:
-    class privateXMLReaderFileV3;
-    privateXMLReaderFileV3 * const mPriv;
-};
+class XMLReaderFileVx;
 
 class YARP_robotinterface_API XMLReader
 {
