+++ src/libYARP_robotinterface/src/yarp/robotinterface/XMLReaderV1.cpp
@@ -6,12 +6,14 @@
  * BSD-3-Clause license. See the accompanying LICENSE file for details.
  */
 
+
 #include "XMLReader.h"
 #include "Action.h"
 #include "Device.h"
 #include "Param.h"
 #include "Robot.h"
 #include "Types.h"
+#include "impl/XMLReaderFileVx.h"
 
 #include <yarp/os/LogStream.h>
 #include <yarp/os/Network.h>
