+++ src/libYARP_robotinterface/src/yarp/robotinterface/XMLReaderV3.cpp
@@ -12,6 +12,7 @@
 #include "Param.h"
 #include "Robot.h"
 #include "Types.h"
+#include "impl/XMLReaderFileVx.h"
 
 #include <yarp/os/LogStream.h>
 #include <yarp/os/Network.h>
